spaceinvaders bullet send to other console
spaceinvaders now has the correct multiplayer implementation.

changed the background of the pause menu and start text to a variable that can be set (borin has a different background than i do)

the screen wasn't resetting from certain choices, e.g. the other console that receives a go back to menu just displayed the menu above the last state of the game, these are fixed